requirements.txt
scikit-image
numba
Pillow